+++ lib/eth/contract/function_input.rb
@@ -16,19 +16,21 @@
 
 # Provides the {Eth} module.
 module Eth
+
   # Provide classes for contract function input.
   class Contract::FunctionInput
-    attr_accessor :type, :name
+    attr_accessor :type, :raw_type, :name
 
     # Constructor of the {Eth::Contract::FunctionInput} class.
     #
     # @param data [Hash] contract abi data.
     def initialize(data)
+      @raw_type = data["type"]
       @type = Eth::Abi::Type.parse(data["type"])
       @name = data["name"]
     end
 
-    # Returns types like uint256
+    # Returns complete types with subtypes, e.g., `uint256`.
     def type
       @type.base_type + @type.sub_type
     end